setup.cfg: restrict upload to official pypi
According to zest's documentation:

Note that since version 3.15, zest.releaser also looks for this
information in the setup.cfg if your package has that file. One way to
use this, is to restrict the servers that zest.releaser will ask you to
upload to.

This will prevent the package to be uploaded on private repositories.